Create Weather App Project
This project is on;
Facade ventures is a multinational brick and mortar store that has stores with large market shares in the following cities: Lusaka (Zambia), Johanessburg (South Africa), Lagos (Nigeria), Accra (Ghana), and Nairobi (Kenya). It has large market shares in each of these cities because of the fast delivery times of the logistics arm of the business. As of Q4 of 2020, the revenue reported by the company was over $5.6 M.
Last year, the business developers at Facade noticed that delivery times can be even faster and made more efficient if the delivery drivers (who deliver across these cities, mostly through shipping and road delivery) can have a way to get weather information that includes;

Human perception of what the weather feels like.
The minimum and maximum temperature in that city,
The speed of the wind in that city,
How cloudy the sky is.
How humid the weather is.
Atmospheric pressure on the ground level.